requirements.txt
requests
SQLAlchemy>=2.0
python-dotenv
